refactor: Remove duplicate factory implementations and simplify codebase
- Delete CommandFactory.java (363 lines) - fully replaced by EnhancedCommandFactory
- Delete AbstractFactory.java (261 lines) - over-generic with unclear value
- Remove deprecated methods with hardcoded default arguments
- Consolidate factory functionality into single EnhancedCommandFactory
- Update all factory references across codebase
- Maintain backward compatibility with existing API
- All 224 tests passing

Total reduction: 624 lines of code
Improves maintainability and reduces duplication
